Day12 k線圖教教我/requirements.txt
streamlit>=1.36.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
openai>=1.35.0
python-dotenv>=1.0.1
